[BE]Changing the addIngredientToRecipe method to add and set simultaneously

+++ src/main/java/com/academy/recipes/spring_course/business/dtos/IngredientRecipeDto.java
@@ -16,6 +16,6 @@ public class IngredientRecipeDto {
 	String recipeName;
 	String ingredientName;
 	String description;
-	int quantity;
+	Float quantity;
 	UnitMeasures unitMeasure;
 }
